refactor: store -> repository 네이밍 변경 및 패키지 위치 변경
