+++ app/src/main/java/com/programmers/order/repository/VoucherRepository.java
@@ -1,10 +1,10 @@
-package com.programmers.order.manager.store;
+package com.programmers.order.repository;
 
 import java.util.List;
 
 import com.programmers.order.domain.Voucher;
 
-public interface VoucherStoreManager {
+public interface VoucherRepository {
 	Voucher saveVoucher(Voucher voucher);
 
 	List<Voucher> getVouchers();
